pyproject.toml
[project]
name = "cc-zol"
version = "1.0.0"
description = "Claude Code CLI with email verification"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "uvicorn>=0.34.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "websockets>=13.0",
    "python-telegram-bot>=21.0",
    "pydantic-settings>=2.12.0",
    "aiolimiter>=1.2.1",
    "openai>=2.16.0",
    "click>=8.0.0",
    "motor>=3.0.0",
]

[project.scripts]
cc-zol = "zol.main:cli"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["zol", "auth", "api", "cli", "config", "messaging", "providers"]
include = ["models.json"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
